feature(Card): Improve Card and config
