+++ src/app.html
@@ -3,7 +3,12 @@
     <head>
         <meta charset="utf-8" />
         <link rel="icon" href="/favicon.png" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1, user-scalable=no"
+        />
+        <meta name="theme-color" content="#66aa55" />
+        <link rel="manifest" href="/manifest.json" />
         %svelte.head%
     </head>
     <body>
