--- src/components/Hello.svelte
@@ -1,1 +1,0 @@
-<h1>Hello!</h1>
