+++ static/manifest.json
@@ -1,0 +1,8 @@
+{
+    "short_name": "Swipe for Future",
+    "name": "Swipe for Future - Solving the climate crisis, one swipe at a time 🌍🔥",
+    "start_url": "/",
+    "display": "standalone",
+    "theme_color": "#66aa55",
+    "background_color": "#ffffff"
+}
